new table performers_personal_data and new registration

+++ handler_performer/performer.py
@@ -17,8 +17,9 @@ from states import performer_states
 
 class PerformerMain:
     @staticmethod
-    async def hi_performer(callback: types.CallbackQuery):
+    async def hi_performer(callback: types.CallbackQuery, state: FSMContext):
         performer = await performers_get.performer_select(callback.from_user.id)
+        performer_p_d = await performers_get.performer_select_personal_data(callback.from_user.id)
         if performer is None:
             await bot.delete_message(callback.from_user.id, callback.message.message_id)
             keyboard = types.ReplyKeyboardMarkup(resize_keyboard=True, row_width=1)
@@ -29,7 +30,7 @@ class PerformerMain:
                                    f"Поделитесь с нами вашим номером телефона!\n",
                                    reply_markup=keyboard)
             await performer_states.PerformerPhone.phone.set()
-        elif performer.ban == 0:
+        if not performer.ban and performer_p_d:
             await bot.delete_message(callback.from_user.id, callback.message.message_id)
             await performer_states.PerformerStart.performer_menu.set()
             await bot.send_message(callback.from_user.id,
@@ -39,12 +40,23 @@ class PerformerMain:
             await bot.send_message(callback.from_user.id,
                                    f"{markup_performer.text_menu(len(orders), len(orders_loading))}",
                                    reply_markup=markup_performer.main_menu())
-        else:
+        if performer.ban == 1:
             await bot.delete_message(callback.from_user.id, callback.message.message_id)
             await bot.send_message(callback.from_user.id, "Вы заблокированы! Обратитесь в техподдержку!")
+        if performer_p_d is None:
+            await performer_states.PerformerStart.info_about_performer.set()
+            await bot.send_message(callback.from_user.id,
+                                   f"{callback.from_user.first_name} Спасибо что пользуетесь нашим ботом!\n"
+                                   f"Теперь пройдите короткую регистрацию")
+            async with state.proxy() as data:
+                data["list_info"] = []
+            await bot.send_message(callback.from_user.id,
+                                   "Введите Ваше реальное Имя\n"
+                                   "Вводить только на русском языке.\n",
+                                   reply_markup=markup_performer.markup_clean)
 
     @staticmethod
-    async def phone(message: types.Message):
+    async def phone(message: types.Message, state: FSMContext):
         if message.contact.user_id == message.from_user.id:
             res = message.contact.phone_number[-10:]
             await performers_set.performer_add(message.from_user.id,
@@ -52,14 +64,49 @@ class PerformerMain:
                                                f'+7{res}',
                                                message.from_user.first_name,
                                                message.from_user.last_name)
-            await performer_states.PerformerStart.performer_menu.set()
+            await performer_states.PerformerStart.info_about_performer.set()
             await bot.send_message(message.from_user.id,
-                                   f"{message.from_user.first_name} Спасибо что пользуетесь нашим ботом!",
-                                   reply_markup=markup_performer.main_menu())
+                                   f"{message.from_user.first_name} Спасибо что пользуетесь нашим ботом!\n"
+                                   f"Теперь пройдите короткую регистрацию")
+            async with state.proxy() as data:
+                data["list_info"] = []
+            await bot.send_message(message.from_user.id,
+                                   "Введите Ваше реальное Имя\n"
+                                   "Вводить только на русском языке.\n",
+                                   reply_markup=markup_performer.markup_clean)
         else:
             await bot.send_message(message.from_user.id,
                                    "Это не ваш номер телефона! \n"
                                    "Нажмите /start чтобы начать заново")
+
+    @staticmethod
+    async def info_about_performer_name(message: types.Message, state: FSMContext):
+        async with state.proxy() as data:
+            if message.photo:
+                data.get("list_info").append(message.photo[2].file_id)
+            if not message.photo:
+                data.get("list_info").append(message.text)
+            if len(data.get("list_info")) == 1:
+                await bot.send_message(message.from_user.id,
+                                       "Введите вашу Фамилию\n"
+                                       "Вводить только на русском языке.\n")
+            if len(data.get("list_info")) == 2:
+                await bot.send_message(message.from_user.id,
+                                       "Сделайте селфи")
+            if len(data.get("list_info")) == 3:
+                await bot.send_message(message.from_user.id,
+                                       "Все получилось!")
+                performer = await performers_get.performer_select(message.from_user.id)
+                await performers_set.performer_add_personal_data(message.from_user.id,
+                                                                 performer.telephone,
+                                                                 data.get("list_info")[0],
+                                                                 data.get("list_info")[1],
+                                                                 data.get("list_info")[2],)
+                await performer_states.PerformerStart.performer_menu.set()
+                await bot.send_message(message.from_user.id,
+                                       "Регистрация завершена!\n"
+                                       "Вы находитесь в главном меню!",
+                                       reply_markup=markup_performer.main_menu())
 
     @staticmethod
     async def main(message: types.Message):
